Enable tess printing in PRC.
A tessellation is now printed in the PRC file by function
nev_print_prc_tess.  The function selectively prints polys, edges and
vertices (depending on the values of -showpoly, -showedge and -showver,
respectively), using colours specified with dedicated options
(-data*col).

Faces are not printed; this is to be fixed.