Add -transform renumber
'-transform renumber' can be used to renumber a tessellation so that no
empty cell is present in the data field, i.e. so that numbers in the
data field go from 1 (or 0 for voids) to the total number of cells.  The
corresponding cell ids are defined in the **cell/*id field (and recorded
in CellId).