Get ready for tree update
The kd-tree information (point cloud and tree) is now recorded in struct
TOPT as pointers to NFCLOUD and NFTREE structures.  The structures
themselves are declared in net_tess_opt as they cannot be stored easily
in struct TOPT (as C++ objects).

This is used to not recompute the tree when no seed has *moved* (changed
position).  When the change is only in terms of seed weight, the tree
does not need to be recomputed.